admin/backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib
python-multipart
